+++ src/binary64/log1p/log1p.c
@@ -313,7 +313,7 @@ double cr_log1p(double x){
     if(__builtin_expect(je<1022, 1)){
       rs.u -= eoff;
     } else {
-      rs.u -= 1021l<<52;
+      rs.u -= (i64)1021<<52;
       static const double sc[] = {0x1p-1, 0x1p-2, 0x1p-3};
       rs.f *= sc[je-1022];
     }
@@ -385,7 +385,7 @@ static double __attribute__((noinline)) as_log1p_refine(double x, double a){
 	t.f = 1 + x;
       dt.u = 0;
     }
-    t.u -= (long)je<<52;
+    t.u -= (int64_t)je<<52;
 
     double t12 = rt[0][j1]*rt[1][j2], t34 = rt[2][j3]*rt[3][j4];
     double th = t12*t34, tl = __builtin_fma(t12,t34,-th);
@@ -394,7 +394,7 @@ static double __attribute__((noinline)) as_log1p_refine(double x, double a){
     double xl, xh = fasttwosum(dh-1, dl, &xl);
     xh = fastsum(xh, xl, sh, sl, &xl);
     if(dt.u){
-      dt.u -= (long)je<<52;
+      dt.u -= (int64_t)je<<52;
       double ddh = th*dt.f, ddl = __builtin_fma(th,dt.f,-ddh) + tl*dt.f;
       xh = fastsum(xh, xl, ddh, ddl, &xl);
     }
